requirements.txt
audioop-lts==0.2.1
mido==1.3.3
numpy==2.2.4
packaging==24.2
pydub==0.25.1
pyfluidsynth==1.3.4
six==1.17.0
